UIの更新: 日本語化とデザインの改善
