+++ lib/models/note.dart
@@ -1,0 +1,7 @@
+class Note {
+  final String title;
+  final String body;
+  final DateTime date;
+
+  Note({required this.title, required this.body, required this.date});
+}
