+++ lib/models/todo.dart
@@ -1,0 +1,6 @@
+class Todo {
+  final String title;
+  bool isDone;
+
+  Todo({required this.title, this.isDone = false});
+}
