+++ lib/screens/add_todo_page.dart
@@ -1,0 +1,46 @@
+import 'package:flutter/material.dart';
+import '../models/todo.dart';
+
+class AddTodoPage extends StatefulWidget {
+  const AddTodoPage({super.key});
+
+  @override
+  State<AddTodoPage> createState() => _AddTodoPageState();
+}
+
+class _AddTodoPageState extends State<AddTodoPage> {
+  final TextEditingController _controller = TextEditingController();
+
+  void _save() {
+    final text = _controller.text.trim();
+    if (text.isEmpty) {
+      Navigator.of(context).pop();
+      return;
+    }
+    final todo = Todo(title: text);
+    Navigator.of(context).pop(todo);
+  }
+
+  @override
+  Widget build(BuildContext context) {
+    return Scaffold(
+      appBar: AppBar(
+        title: const Text('新規Todo'),
+        backgroundColor: Theme.of(context).colorScheme.inversePrimary,
+      ),
+      body: Padding(
+        padding: const EdgeInsets.all(16),
+        child: Column(
+          children: [
+            TextField(
+              controller: _controller,
+              decoration: const InputDecoration(labelText: 'Todo'),
+            ),
+            const SizedBox(height: 16),
+            ElevatedButton(onPressed: _save, child: const Text('保存')),
+          ],
+        ),
+      ),
+    );
+  }
+}
